module Text.Commonmark.ParserCombinators.Prim
Moved everything that accesses `Parser` internals
into a separate module, `Parser` and `ParserState` constructors
aren't exported from it.
